+++ java/src/test/java/io/cucumber/cucumberexpressions/CustomMatchers.java
@@ -8,6 +8,7 @@ import org.hamcrest.core.IsEqual;
 import org.hamcrest.number.IsCloseTo;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -31,6 +32,8 @@ public class CustomMatchers {
         @Override
         public boolean matches(Object actual) {
             if(actual instanceof BigDecimal) {
+                return new IsEqual(this.expectedValue).matches(actual.toString());
+            } else if(actual instanceof BigInteger) {
                 return new IsEqual(this.expectedValue).matches(actual.toString());
             } else if(actual instanceof Double || actual instanceof Float) {
                 return new IsCloseTo(((Double)this.expectedValue), 0.0001).matches(((Number)actual).doubleValue());